AT-6113 fixinf format, folders names and dependicies setup
