--- .secrets.baseline
@@ -1,248 +1,0 @@
-{
-  "version": "1.0.3",
-  "plugins_used": [
-    {
-      "name": "Base64HighEntropyString",
-      "limit": 4.5
-    },
-    {
-      "name": "BasicAuthDetector"
-    },
-    {
-      "name": "HexHighEntropyString",
-      "limit": 3
-    },
-    {
-      "name": "KeywordDetector",
-      "keyword_exclude": ""
-    },
-    {
-      "name": "PrivateKeyDetector"
-    }
-  ],
-  "filters_used": [
-    {
-      "path": "detect_secrets.filters.allowlist.is_line_allowlisted"
-    },
-    {
-      "path": "detect_secrets.filters.common.is_baseline_file",
-      "filename": ".secrets.baseline"
-    },
-    {
-      "path": "detect_secrets.filters.common.is_ignored_due_to_verification_policies",
-      "min_level": 2
-    },
-    {
-      "path": "detect_secrets.filters.heuristic.is_indirect_reference"
-    },
-    {
-      "path": "detect_secrets.filters.heuristic.is_likely_id_string"
-    },
-    {
-      "path": "detect_secrets.filters.heuristic.is_potential_uuid"
-    },
-    {
-      "path": "detect_secrets.filters.heuristic.is_prefixed_with_dollar_sign"
-    },
-    {
-      "path": "detect_secrets.filters.heuristic.is_sequential_string"
-    },
-    {
-      "path": "detect_secrets.filters.heuristic.is_templated_secret"
-    },
-    {
-      "path": "detect_secrets.filters.regex.should_exclude_file",
-      "pattern": [
-        "^.secrets.baseline$|^.*pgsslrootcert.yml$"
-      ]
-    }
-  ],
-  "results": {
-    ".bandit_config": [
-      {
-        "type": "Secret Keyword",
-        "filename": ".bandit_config",
-        "hashed_secret": "789b9f59f0782902be35cc09a4d97fced60a2c5d",
-        "is_verified": false,
-        "line_number": 42
-      }
-    ],
-    ".github/workflows/sonar.yml": [
-      {
-        "type": "Secret Keyword",
-        "filename": ".github/workflows/sonar.yml",
-        "hashed_secret": "afc848c316af1a89d49826c5ae9d00ed769415f3",
-        "is_verified": false,
-        "line_number": 28
-      }
-    ],
-    "PortfolioProvision.md": [
-      {
-        "type": "Secret Keyword",
-        "filename": "PortfolioProvision.md",
-        "hashed_secret": "3442496b96dd01591a8cd44b1eec1368ab728aba",
-        "is_verified": false,
-        "line_number": 26
-      }
-    ],
-    "README.md": [
-      {
-        "type": "Secret Keyword",
-        "filename": "README.md",
-        "hashed_secret": "d141ce86b0584abb29ee7c24af9afb1e3d871f04",
-        "is_verified": false,
-        "line_number": 193
-      }
-    ],
-    "alembic.ini": [
-      {
-        "type": "Basic Auth Credentials",
-        "filename": "alembic.ini",
-        "hashed_secret": "9d4e1e23bd5b727046a9e3b4b7db57bd8d6ee684",
-        "is_verified": false,
-        "line_number": 38
-      }
-    ],
-    "alembic/versions/e0c6eb21771f_reset_migrations_with_new_schema.py": [
-      {
-        "type": "Hex High Entropy String",
-        "filename": "alembic/versions/e0c6eb21771f_reset_migrations_with_new_schema.py",
-        "hashed_secret": "999a22300a564f9d2bdca555c2170465fd760ae3",
-        "is_verified": false,
-        "line_number": 13
-      }
-    ],
-    "alembic/versions/fda6bd7e1b65_clin_delete_cascade.py": [
-      {
-        "type": "Hex High Entropy String",
-        "filename": "alembic/versions/fda6bd7e1b65_clin_delete_cascade.py",
-        "hashed_secret": "61d8937fb12b982e07b933c083d9014c34159723",
-        "is_verified": false,
-        "line_number": 13
-      },
-      {
-        "type": "Hex High Entropy String",
-        "filename": "alembic/versions/fda6bd7e1b65_clin_delete_cascade.py",
-        "hashed_secret": "999a22300a564f9d2bdca555c2170465fd760ae3",
-        "is_verified": false,
-        "line_number": 14
-      }
-    ],
-    "ansible/README.md": [
-      {
-        "type": "Secret Keyword",
-        "filename": "ansible/README.md",
-        "hashed_secret": "e5e9fa1ba31ecd1ae84f75caaa474f3a663f05f4",
-        "is_verified": false,
-        "line_number": 52
-      },
-      {
-        "type": "Secret Keyword",
-        "filename": "ansible/README.md",
-        "hashed_secret": "5718bce97710e6be87ea160b36eaefb5032857d3",
-        "is_verified": false,
-        "line_number": 124
-      }
-    ],
-    "ansible/roles/db/tasks/main.yml": [
-      {
-        "type": "Secret Keyword",
-        "filename": "ansible/roles/db/tasks/main.yml",
-        "hashed_secret": "d2e2ab0f407e4ee3cf2ab87d61c31b25a74085e5",
-        "is_verified": false,
-        "line_number": 22
-      }
-    ],
-    "atat.ini.example": [
-      {
-        "type": "Secret Keyword",
-        "filename": "atat.ini.example",
-        "hashed_secret": "abcdb568713c255c81376829da20004ba9463fd3",
-        "is_verified": false,
-        "line_number": 2
-      }
-    ],
-    "atat/domain/csp/cloud/mock_cloud_provider.py": [
-      {
-        "type": "Secret Keyword",
-        "filename": "atat/domain/csp/cloud/mock_cloud_provider.py",
-        "hashed_secret": "a94a8fe5ccb19ba61c4c0873d391e987982fbbd3",
-        "is_verified": false,
-        "line_number": 137
-      }
-    ],
-    "config/base.ini": [
-      {
-        "type": "Secret Keyword",
-        "filename": "config/base.ini",
-        "hashed_secret": "afc848c316af1a89d49826c5ae9d00ed769415f3",
-        "is_verified": false,
-        "line_number": 54
-      }
-    ],
-    "docs/SAMLIDP.md": [
-      {
-        "type": "Secret Keyword",
-        "filename": "docs/SAMLIDP.md",
-        "hashed_secret": "1c1ded6cd9b5d5fe2bf39701732f748da346763b",
-        "is_verified": false,
-        "line_number": 13
-      }
-    ],
-    "docs/WindowsSubsystemLinux.md": [
-      {
-        "type": "Secret Keyword",
-        "filename": "docs/WindowsSubsystemLinux.md",
-        "hashed_secret": "afc848c316af1a89d49826c5ae9d00ed769415f3",
-        "is_verified": false,
-        "line_number": 104
-      }
-    ],
-    "ops/app.tfvars.json.template": [
-      {
-        "type": "Secret Keyword",
-        "filename": "ops/app.tfvars.json.template",
-        "hashed_secret": "1f5e25be9b575e9f5d39c82dfd1d9f4d73f1975c",
-        "is_verified": false,
-        "line_number": 13
-      }
-    ],
-    "poetry.lock": [
-      {
-        "type": "Hex High Entropy String",
-        "filename": "poetry.lock",
-        "hashed_secret": "58b2ce6610a2e13361a9ce44ed69ce2e49039fed",
-        "is_verified": false,
-        "line_number": 1498
-      }
-    ],
-    "tests/routes/task_orders/test_new.py": [
-      {
-        "type": "Hex High Entropy String",
-        "filename": "tests/routes/task_orders/test_new.py",
-        "hashed_secret": "e4f14805dfd1e6af030359090c535e149e6b4207",
-        "is_verified": false,
-        "line_number": 41
-      }
-    ],
-    "tests/test_access.py": [
-      {
-        "type": "Hex High Entropy String",
-        "filename": "tests/test_access.py",
-        "hashed_secret": "e4f14805dfd1e6af030359090c535e149e6b4207",
-        "is_verified": false,
-        "line_number": 637
-      }
-    ],
-    "uitests/README.md": [
-      {
-        "type": "Secret Keyword",
-        "filename": "uitests/README.md",
-        "hashed_secret": "81b127e2222d9bfc4609053faec85300f7525463",
-        "is_verified": false,
-        "line_number": 58
-      }
-    ]
-  },
-  "generated_at": "2021-03-30T18:10:37Z"
-}
