--- uitests/TestFramework/Driver_Notes/LICENSE:Zone.Identifier
@@ -1,3 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=C:\Users\EdwardHoward(ASG)\Downloads\edgedriver_win64 (3).zip
