--- uitests/TestFramework/IEDriverServer.exe:Zone.Identifier
@@ -1,3 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=C:\Users\EdwardHoward(ASG)\Downloads\IEDriverServer_x64_3.150.1.zip
