--- uitests/TestFramework/chromedriver.exe:Zone.Identifier
@@ -1,3 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=C:\Users\EdwardHoward(ASG)\Downloads\chromedriver_win32 (4).zip
