--- uitests/TestFramework/geckodriver.exe:Zone.Identifier
@@ -1,3 +1,0 @@
-[ZoneTransfer]
-ZoneId=3
-ReferrerUrl=C:\Users\EdwardHoward(ASG)\Downloads\geckodriver-v0.29.0-win64.zip
