+++ uitests/framework/utilities/custom_logger.py
@@ -1,0 +1,14 @@
+import logging
+
+
+class LogGen:
+    @staticmethod
+    def loggen():
+        logging.basicConfig(
+            format="%(asctime)s %(message)s",
+            datefmt="%m/%d/%Y %I:%M:%S %p",
+            filename="./Logs/example.log",
+            level=logging.INFO,
+        )
+        logger = logging.getLogger()
+        return logger
